+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Encode hash in base64 in order to make needed URL prefix shorter at same "guessability".
   → This causes the maximum prefix length to go down from 128 to 64 (4 bit per char → 8 bit per char)
 * `list`: Add `--with-size`-switch to print file sizes.
+* `list`: Add `--filenames`-switch to print filenames instead of full urls in listing.
 
 ## v0.3.1 (2020-10-03)