+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # Changelog for [`asfa`](https://github.com/obreitwi/asfa)
 
-## v0.3.2-pre (under development)
+## v0.4.0 (2020-10-14)
 
 * Add `private_key_file` to auth-option in order to specfiy private key file directly
 * Add `--verbose` argument that increases loglevel